Add coefficient of variance (CV) limit option (#11)
This PR adds support for normalized standard deviation (also known as coefficient of variance, or CV) analysis. This is a more standardized and likely user-friendly way to look at how much values vary. Settling on a standard limit of, say, 20% might be reasonable (or maybe less?).

This PR adds the new analysis to CI (with a limit of 5%). It also includes some refactoring to improve stability of its own code. It also fixes a bug in `double` to `BigRational` (`real`) conversion.